[mcu] Add a dirty (?) fix for re-aliging to an open page.
Previously, if you opened a page, waited a bit (?) in an inactive state
and then requested aligning with the already opened page, the MCU would
precharge and open a page.

This fix prevents that, but at the cost of adding a cycle of latency when
aligning with a non-open page.